+++ Sem. 04/Pract.04/task16.cpp
@@ -10,22 +10,7 @@ int getRemainder(int first, int second) {
     return result;
 }
 
-int getDivision(int first, int second) {
-    int result = 0;
-    if (first >= second) {
-        int remainder = (first % second);
-        first -= remainder;
-
-        while (first >= second) {
-            first -= second;
-            result++;
-        }
-    }
-    return result;
-}
-
 int main() {
     std::cout << getRemainder(17, 3) << '\n';
-    std::cout << getDivision(12, 2) << '\n';
     return 0;
 }
